VMS-40962: Support large requests for System Database restore endpoint
(cherry picked from commit 11b52cc13bca8b8feb064fc34dc1328a1106f4d7)

+++ vms/libs/nx_vms_common/src/network/tcp_connection_priv.h
@@ -129,6 +129,6 @@ private:
     nx::Mutex sendDataMutex;
     QByteArray interleavedMessageData;
     size_t interleavedMessageDataPos;
-    size_t currentRequestSize;
+    qint64 currentRequestSize;
     std::optional<std::chrono::time_point<std::chrono::steady_clock>> lastSendTime;
 };
